+++ CHANGELOG.md
@@ -14,11 +14,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 -   Map filters placed in sidebar component
+-   Clear Filters and Filter Sites buttons moved back from filter to map component
 
 ### Fixed
 
 -   Width of map container on mobile fits to screen size
 -   Event list is reset when "Clear Filters" is clicked and active styling for previously selected event types removed
+-   Fixed map filter overlap with results panel and footer
 
 ## [v0.5.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.5.0) - 2021-06-10